+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG
 
-* **2024.02.15    Current**
+* **2024.02.18    Current**
+  * `SSH2` proxy supported in any position of a `CHAIN`
   * `gui-warp.app`: on macOS deiconify from Dock; `LICENSE` file added to the application
-  * `ts-warp.c`: Drop SSH2 proxy connection if it's in a mid of a `CHAIN`. `Single SSH2-proxy` and
-    `SSH2-proxy-as-a-target` of a `CHAIN` are supported
+  * ~~`ts-warp.c`: Drop SSH2 proxy connection if it's in a mid of a `CHAIN`. `Single SSH2-proxy` and~~
+    ~~`SSH2-proxy-as-a-target` of a `CHAIN` are supported~~
   * `http.c`: `http_client_request()` tolerate fail on `send()` and `recv()`
   * `ts-warp.c`: `NIT lookup` lifted up the code to work for all proxies
   * `ts-warp.c`: minor logging bugfix, `README.md` updates